Splits `coordinator` into sub-files
