+++ lib/application/pages/home/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:memo/application/coordinator.dart';
+import 'package:memo/application/coordinator/routes_coordinator.dart';
 
 enum HomeBottomTab { study, progress }
